backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
tensorflow==2.16.1
opencv-python-headless==4.9.0.80
pillow==10.3.0
numpy==1.26.4
scikit-learn==1.4.2
pandas==2.2.2
matplotlib==3.9.0
python-dotenv==1.0.1
gdown==6.0.0
httpx==0.27.0
pydantic==2.7.1
